+++ test/clojure/clojurewerkz/ogre/suite/dedup_test.clj
@@ -161,3 +161,22 @@
               (select (column :values))
               (unfold)
               (dedup)))
+
+(defn get_g_V_both_group_by_byXout_dedup_foldX_unfold_selectXvaluesX_unfold_out_order_byXnameX_limitX1X_valuesXnameX
+  "g.V().both().
+     group().by().by(__.out().dedup().fold()).
+     unfold().
+     select(Column.values).
+     unfold().
+     out().order().by('name').limit(1).values('name')"
+  [g]
+  (traverse g (V)
+            (both)
+            (group) (by) (by (__ (out) (dedup) (fold)))
+            (unfold)
+            (select (column :values))
+            (unfold)
+            (out)
+            (order) (by :name)
+            (limit 1)
+            (values :name)))