+++ changelog.md
@@ -1,7 +1,7 @@
 ## Changes in Ogre 3.4.8.0
 
 * Bump to Apache TinkerPop 3.4.8
-* Removed `with-remote` function which used some deprecated functions in TinkerPop and allowed the `traversal` function to take a `RemoteConnection` directly.
+* Allowed the `traversal` function to take a `RemoteConnection` directly which is more in line with TinkerPop `AnonymousTraversalSource` method of construction - the `with-remote`standalone function is no longer recommended. 
 
 ## Changes in Ogre 3.4.7.0